added dashboard-card module
