+++ src/app/core/dashboard/dashboard-card/dashboard-card.component.css
@@ -1,0 +1,93 @@
+.dashboard-card {
+  position: absolute;
+  top: 0.5rem;
+  left: 0.5rem;
+  right: 0.5rem;
+  bottom: 1.5rem;
+}
+.dashboard-card-content {
+  text-align: center;
+}
+.mat-card-header {
+  justify-content: flex-end;
+}
+#chart {
+  max-width: 350px;
+}
+
+.mat-table {
+  width: 100%;
+}
+
+td.progress-td {
+  width: 2rem;
+  max-width: 3.5rem;
+}
+
+.mat-row {
+  height: fit-content;
+}
+
+mat-progress-bar {
+  height: 0.5rem;
+  width: 3rem;
+  border-radius: 1.25em;
+}
+
+.mat-cell {
+  padding-left: 0.25rem;
+  text-align: left;
+}
+
+.mat-cell:first-child {
+  width: 0.5rem;
+  max-width: 3rem;
+  padding: 0 0.25rem;
+}
+
+.mat-cell:last-child {
+  padding-right: 0.25rem;
+  text-align: right;
+}
+
+::ng-deep .mat-progress-bar-fill::after {
+  background-color: #00E396;
+  border-radius: 1.25em;
+}
+
+::ng-deep .mat-progress-bar-buffer {
+  background: #f2f2f2;;
+}
+
+.green-progress-bar > ::ng-deep .mat-progress-bar-fill::after {
+  background-color: #00E396;
+}
+
+.yellow-progress-bar > ::ng-deep .mat-progress-bar-fill::after {
+  background-color: #e7f436;
+}
+
+.orange-progress-bar > ::ng-deep .mat-progress-bar-fill::after {
+  background-color: #FEB019;
+}
+
+.red-progress-bar > ::ng-deep .mat-progress-bar-fill::after {
+  background-color: #FF4560;
+}
+.current-user-rank-row {
+  background-color: #546E7A;
+  /*color: white;*/
+  border-radius: 1.25em;
+}
+.current-user-rank-row>td {
+  color: white;
+}
+.current-user-rank-row>td:first-child {
+  border-top-left-radius: 1.25em;
+  border-bottom-left-radius: 1.25em;
+}
+.current-user-rank-row>td:last-child {
+  border-top-right-radius: 1.25em;
+  border-bottom-right-radius: 1.25em;
+}
+
