+++ src/app/core/dashboard/dashboard-card/dashboard-card.module.ts
@@ -1,0 +1,24 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { DashboardCardComponent } from './dashboard-card.component';
+import {MatCardModule} from '@angular/material/card';
+import {MatTableModule} from '@angular/material/table';
+import {MatProgressBarModule} from '@angular/material/progress-bar';
+import {NgApexchartsModule} from 'ng-apexcharts';
+
+
+
+@NgModule({
+  declarations: [DashboardCardComponent],
+  exports: [
+    DashboardCardComponent,
+  ],
+  imports: [
+    CommonModule,
+    MatCardModule,
+    MatTableModule,
+    MatProgressBarModule,
+    NgApexchartsModule
+  ]
+})
+export class DashboardCardModule { }
