+++ src/app/core/dashboard/dashboard.component.html
@@ -14,40 +14,13 @@
       </mat-card>
     </mat-grid-tile>
     <mat-grid-tile [colspan]="1" [rowspan]="1">
-      <mat-card class="dashboard-card">
-        <mat-card-header>
-          <mat-card-title>
-            ППР
-          </mat-card-title>
-        </mat-card-header>
-        <mat-card-content class="dashboard-card-content">
-          <app-ppr-dashboard></app-ppr-dashboard>
-        </mat-card-content>
-      </mat-card>
+      <app-ppr-dashboard></app-ppr-dashboard>
     </mat-grid-tile>
     <mat-grid-tile [colspan]="1" [rowspan]="1">
-      <mat-card class="dashboard-card">
-        <mat-card-header>
-          <mat-card-title>
-            Індивідуальний план
-          </mat-card-title>
-        </mat-card-header>
-        <mat-card-content class="dashboard-card-content">
-          <app-individual-dashboard></app-individual-dashboard>
-        </mat-card-content>
-      </mat-card>
+      <app-individual-dashboard></app-individual-dashboard>
     </mat-grid-tile>
     <mat-grid-tile [colspan]="1" [rowspan]="1">
-      <mat-card class="dashboard-card">
-        <mat-card-header>
-          <mat-card-title>
-            Партнер 3+
-          </mat-card-title>
-        </mat-card-header>
-        <mat-card-content class="dashboard-card-content">
-          <app-partner-dashboard></app-partner-dashboard>
-        </mat-card-content>
-      </mat-card>
+      <app-partner-dashboard></app-partner-dashboard>
     </mat-grid-tile>
   </mat-grid-list>
 </div>
