+++ src/app/core/dashboard/individual-dashboard/individual-dashboard.component.html
@@ -1,0 +1,6 @@
+<app-dashboard-card
+  [dashboardCardName]="dashboardCardName"
+  [userCode]="userCode"
+  [radialBarSeries]="radialBarSeries"
+  [employees]="employees"
+></app-dashboard-card>
