+++ src/app/core/dashboard/individual-dashboard/individual-dashboard.component.ts
@@ -1,4 +1,5 @@
 import {Component, OnInit} from '@angular/core';
+import {Employee} from '../dashboard-card/dashboard-card.component';
 
 @Component({
   selector: 'app-individual-dashboard',
@@ -6,8 +7,22 @@ import {Component, OnInit} from '@angular/core';
   styleUrls: ['./individual-dashboard.component.css']
 })
 export class IndividualDashboardComponent implements OnInit {
+  dashboardCardName: string;
+  userCode: number;
+  radialBarSeries: number[];
+  employees: Employee[];
 
-  constructor() {}
+  constructor() {
+    this.dashboardCardName = 'Індивідуальний план';
+    this.userCode = 12345;
+    this.radialBarSeries = [12, 56];
+    this.employees = [
+      {name: 'Hot', level: 1.05},
+      {name: 'Petriv', level: 0.78},
+      {name: 'Shkvarok', level: 0.56, userCode: 12345},
+      {name: 'Petrivkyi', level: 0.32}
+    ];
+  }
 
   ngOnInit(): void {
   }
