--- src/app/core/dashboard/partner-dashboard/partner-dashboard.component.css
@@ -1,81 +1,0 @@
-#chart {
-  /*padding: 0;*/
-  max-width: 350px;
-  /*margin: 35px auto;*/
-}
-
-.mat-table {
-  width: 100%;
-}
-
-td.progress-td {
-  width: 2rem;
-  max-width: 3.5rem;
-}
-
-.mat-row {
-  height: fit-content;
-}
-
-mat-progress-bar {
-  height: 0.5rem;
-  width: 3rem;
-  border-radius: 1.25em;
-}
-
-.mat-cell {
-  padding-left: 0.25rem;
-  text-align: left;
-}
-
-.mat-cell:first-child {
-  width: 0.5rem;
-  max-width: 3rem;
-  padding: 0 0.25rem;
-}
-
-.mat-cell:last-child {
-  padding-right: 0.25rem;
-  text-align: right;
-}
-
-::ng-deep .mat-progress-bar-fill::after {
-  background-color: #00E396;
-  border-radius: 1.25em;
-}
-
-::ng-deep .mat-progress-bar-buffer {
-  background: #f2f2f2;;
-}
-
-.green-progress-bar > ::ng-deep .mat-progress-bar-fill::after {
-  background-color: #00E396;
-}
-
-.yellow-progress-bar > ::ng-deep .mat-progress-bar-fill::after {
-  background-color: #e7f436;
-}
-
-.orange-progress-bar > ::ng-deep .mat-progress-bar-fill::after {
-  background-color: #FEB019;
-}
-
-.red-progress-bar > ::ng-deep .mat-progress-bar-fill::after {
-  background-color: #FF4560;
-}
-.current-user-rank-row {
-  background-color: #546E7A;
-  /*color: white;*/
-  border-radius: 1.25em;
-}
-.current-user-rank-row>td {
-  color: white;
-}
-.current-user-rank-row>td:first-child {
-  border-top-left-radius: 1.25em;
-  border-bottom-left-radius: 1.25em;
-}
-.current-user-rank-row>td:last-child {
-  border-top-right-radius: 1.25em;
-  border-bottom-right-radius: 1.25em;
-}
