+++ src/app/core/dashboard/partner-dashboard/partner-dashboard.component.html
@@ -1,38 +1,7 @@
-<table mat-table [dataSource]="employees">
-  <ng-container matColumnDef="index">
-    <td mat-cell *matCellDef="let el; let i = index;">
-      {{i + 1}}
-    </td>
-  </ng-container>
-  <ng-container matColumnDef="name">
-    <td mat-cell *matCellDef="let el">
-      {{el['name']}}
-    </td>
-  </ng-container>
-  <ng-container matColumnDef="progressBar">
-    <td mat-cell class="progress-td" *matCellDef="let el">
-      <mat-progress-bar
-        mode="determinate"
-        [class]="updateProgressBarColor(el['level'])"
-        [value]="el['level'] * 100"
-      ></mat-progress-bar>
-    </td>
-  </ng-container>
-  <ng-container matColumnDef="level">
-    <td mat-cell class="progress-td" *matCellDef="let el">
-      {{el['level'] | percent}}
-    </td>
-  </ng-container>
-  <tr mat-row *matRowDef="let row; columns: columns" [ngClass]="updateRankTableRowColor(row)"></tr>
-</table>
-<div id="chart">
-  <apx-chart
-    [chart]="indicatorChartOptions.chart"
-    [fill]="indicatorChartOptions.fill"
-    [labels]="indicatorChartOptions.labels"
-    [plotOptions]="indicatorChartOptions.plotOptions"
-    [series]="indicatorChartOptions.series"
-    [stroke]="indicatorChartOptions.stroke"
-    [theme]="indicatorChartOptions.theme"
-  ></apx-chart>
-</div>
+<app-dashboard-card
+  [dashboardCardName]="dashboardCardName"
+  [userCode]="userCode"
+  [radialBarSeries]="radialBarSeries"
+  [employees]="employees"
+></app-dashboard-card>
+
