+++ src/app/core/dashboard/ppr-dashboard/ppr-dashboard.component.html
@@ -1,1 +1,6 @@
-
+<app-dashboard-card
+  [dashboardCardName]="dashboardCardName"
+  [userCode]="userCode"
+  [radialBarSeries]="radialBarSeries"
+  [employees]="employees"
+></app-dashboard-card>
